requirements.txt
aws-cdk-lib==2.91.0
constructs>=10.0.0,<11.0.0
boto3==1.34.44